example/requirements.txt
-e ../
Django==1.11.6
djangorestframework==3.7.1
djangorestframework-csv==2.0.0
